스태미너 클래스 수정 및 스테미너 관련 아이템 적용
회복, 소모 매개변수로 들어온 값에 따라서 사용되게 변경
스테미너 회복, 획득 증가량 아이템 적용

+++ Assets/WeaponAsset/Scripts/DataStore.cs
@@ -296,7 +296,7 @@ public class DataStore : MonoBehaviourSingleton<DataStore>
         float chargeTimeMax;
         float criticalChance;
         float damage;
-        float staminaConsumption;
+        int staminaConsumption;
         float cooldown;
         int ammoCapacity;
         float range;
@@ -378,7 +378,7 @@ public class DataStore : MonoBehaviourSingleton<DataStore>
             tempWeaponInfos[i].damage = damage;
             //Debug.Log(damage);
 
-            float.TryParse(weaponDatas[i]["staminaConsumption"].ToString(), out staminaConsumption);
+            int.TryParse(weaponDatas[i]["staminaConsumption"].ToString(), out staminaConsumption);
             tempWeaponInfos[i].staminaConsumption = staminaConsumption;
             //Debug.Log(staminaConsumption);
